refactor: move brace pattern validation to separate file
